# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QRX API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: donationPaid
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentListResponse"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment (public, no login required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/track:
    get:
      operationId: trackAppointment
      tags: [appointments]
      summary: Track appointment by phone number
      parameters:
        - name: phone
          in: query
          required: true
          schema:
            type: string
        - name: doctorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Appointment tracking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentTrackResponse"
  /appointments/{id}/add-to-queue:
    post:
      operationId: addToQueue
      tags: [appointments]
      summary: Add appointment to today's queue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Added to queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"

  # ---- ASSISTANTS ----
...
components:
  schemas:
...
    Appointment:
      type: object
      required: [id, doctorId, patientName, patientPhone, appointmentDate, serialNo, status]
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        doctorName:
          type: ["string", "null"]
        patientName:
          type: string
        patientPhone:
          type: string
        patientAge:
          type: ["integer", "null"]
        patientGender:
          type: ["string", "null"]
        complaint:
          type: ["string", "null"]
        bp:
          type: ["string", "null"]
        pulse:
          type: ["string", "null"]
        temp:
          type: ["string", "null"]
        weight:
          type: ["string", "null"]
        height:
          type: ["string", "null"]
        hb:
          type: ["string", "null"]
        sugar:
          type: ["string", "null"]
        spo2:
          type: ["string", "null"]
        medicalHistory:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        labReportUrl:
          type: ["string", "null"]
        prescriptionUploadUrl:
          type: ["string", "null"]
        bookingSource:
          type: ["string", "null"]
        appointmentDate:
          type: string
        appointmentTime:
          type: ["string", "null"]
        serialNo:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        donationPaid:
          type: boolean
        donationAmount:
          type: ["integer", "null"]
        donationPaidAt:
          type: ["string", "null"]
        createdAt:
          type: string
    AppointmentListResponse:
      type: object
      required: [appointments, total, page, limit]
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AppointmentInput:
      type: object
      required: [doctorId, patientName, patientPhone, appointmentDate]
      properties:
        doctorId:
          type: integer
        patientName:
          type: string
        patientPhone:
          type: string
        patientAge:
          type: integer
        patientGender:
          type: string
        complaint:
          type: string
        patientEmail:
          type: string
        bp:
          type: string
        pulse:
          type: string
        temp:
          type: string
        weight:
          type: string
        height:
          type: string
        hb:
          type: string
        sugar:
          type: string
        spo2:
          type: string
        medicalHistory:
          type: string
        notes:
          type: string
        labReportUrl:
          type: string
        prescriptionUploadUrl:
          type: string
        bookingSource:
          type: string
        appointmentDate:
          type: string
        appointmentTime:
          type: string
    AppointmentUpdate:
      type: object
      properties:
        status:
          type: string
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        bp:
          type: string
        pulse:
          type: string
        temp:
          type: string
        weight:
          type: string
        height:
          type: string
        hb:
          type: string
        sugar:
          type: string
        spo2:
          type: string
        medicalHistory:
          type: string
        notes:
          type: string
        labReportUrl:
          type: string
        prescriptionUploadUrl:
          type: string
        complaint:
          type: string
...
    AppointmentTrackResponse:
      type: object
      required: [appointments]
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentTrackItem"
    AppointmentTrackItem:
      type: object
      required: [id, serialNo, status, doctorName, appointmentDate]
      properties:
        id:
          type: integer
        serialNo:
          type: integer
        status:
          type: string
        doctorName:
          type: string
        appointmentDate:
          type: string
        currentServingSerial:
          type: ["integer", "null"]
        patientsAhead:
          type: integer
        waitingCount:
          type: ["integer", "null"]

    # QUEUE
    QueueEntry:
      type: object
      required: [id, doctorId, patientName, serialNo, status]
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        appointmentId:
          type: ["integer", "null"]
        patientName:
          type: string
        patientPhone:
          type: ["string", "null"]
        serialNo:
          type: integer
        status:
          type: string
          enum: [waiting, serving, seen, skipped]
        queueDate:
          type: string
        createdAt:
          type: string
...